Missed file to refactor; fix case when fail to save to non-ssh repo
