+++ blueocean-git-pipeline/src/main/java/io/jenkins/blueocean/blueocean_git_pipeline/GitReadSaveService.java
@@ -77,6 +77,14 @@ public class GitReadSaveService extends ScmContentProvider {
 
     @Override
     public String getApiUrl(@Nonnull Item item) {
+        if (item instanceof org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject) {
+            MultiBranchProject<?,?> mbp = (MultiBranchProject<?,?>)item;
+            SCMSource s = mbp.getSCMSources().iterator().next();
+            if (s instanceof GitSCMSource) {
+                String remote = ((GitSCMSource)s).getRemote();
+                return remote;
+            }
+        }
         return null;
     }
 
@@ -197,13 +205,6 @@ public class GitReadSaveService extends ScmContentProvider {
 
     @Override
     public boolean support(@Nonnull Item item) {
-        if (item instanceof org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject) {
-            MultiBranchProject<?,?> mbp = (MultiBranchProject<?,?>)item;
-            SCMSource s = mbp.getSCMSources().iterator().next();
-            if (s instanceof GitSCMSource) {
-                return true;
-            }
-        }
-        return false;
+        return getApiUrl(item) != null;
     }
 }
