+++ blueocean-git-pipeline/src/main/java/io/jenkins/blueocean/ssh/UserKey.java
@@ -1,4 +1,4 @@
-package io.jenkins.blueocean.service.embedded.rest;
+package io.jenkins.blueocean.ssh;
 
 import org.kohsuke.stapler.export.Exported;
 import org.kohsuke.stapler.export.ExportedBean;
