+++ blueocean-git-pipeline/src/main/java/io/jenkins/blueocean/ssh/UserSSHKeyManager.java
@@ -36,7 +36,6 @@ import java.io.IOException;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 
-import io.jenkins.blueocean.service.embedded.rest.UserKey;
 import jenkins.model.Jenkins;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
